requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
ipykernel
python-jose[cryptography]     
psycopg2-binary
bcrypt==4.0.1
passlib[bcrypt]==1.7.4